chore: fmt gibwork
