+++ solagent-adapters/rig/gibwork/.gitignore
@@ -1,0 +1,1 @@
+target/
